# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tenants": {
      "post": {
        "operationId": "TenantsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workspace created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 201,
                    "message": "Workspace created successfully",
                    "data": {
                      "id": "tenant-id",
                      "name": "My Workspace",
                      "domain": "my-workspace"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Create a new workspace",
        "tags": [
          "Tenants"
        ]
      },
      "get": {
        "operationId": "TenantsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Workspaces retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Workspaces retrieved successfully",
                    "data": [
                      {
                        "id": "tenant-id",
                        "name": "My Workspace",
                        "domain": "my-workspace"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "List all workspaces the user belongs to",
        "tags": [
          "Tenants"
        ]
      }
    },
    "/tenants/{id}": {
      "get": {
        "operationId": "TenantsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace details returned",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Workspace details retrieved successfully",
                    "data": {
                      "id": "tenant-id",
                      "name": "My Workspace",
                      "domain": "my-workspace",
                      "isActive": true
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Get workspace details",
        "tags": [
          "Tenants"
        ]
      },
      "patch": {
        "operationId": "TenantsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workspace updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Workspace updated successfully",
                    "data": {
                      "id": "tenant-id",
                      "name": "My Updated Workspace",
                      "domain": "my-updated-workspace"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Update workspace settings (Owner only)",
        "tags": [
          "Tenants"
        ]
      },
      "delete": {
        "operationId": "TenantsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Workspace deleted successfully",
                    "data": null
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Delete workspace (Owner only)",
        "tags": [
          "Tenants"
        ]
      }
    },
    "/tenants/{tenantId}/members/invite": {
      "post": {
        "operationId": "TenantMembersController_inviteMember",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Member invited successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 201,
                    "message": "Member invited successfully",
                    "data": {
                      "id": "invite-id",
                      "email": "newmember@example.com",
                      "roleId": "role-id",
                      "expiresAt": "2026-06-23T00:00:00Z"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Invite a new member to the workspace (Owner only)",
        "tags": [
          "Tenant Members"
        ]
      }
    },
    "/auth/invitations/accept": {
      "post": {
        "operationId": "TenantMembersController_acceptInvitation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invitation accepted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Invitation accepted successfully",
                    "data": {
                      "tenantId": "tenant-id",
                      "userId": "user-id",
                      "isOwner": false
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Accept a workspace invitation",
        "tags": [
          "Tenant Members"
        ]
      }
    },
    "/tenants/{tenantId}/members": {
      "get": {
        "operationId": "TenantMembersController_getMembers",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Members retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Members retrieved successfully",
                    "data": [
                      {
                        "userId": "user-id",
                        "tenantId": "tenant-id",
                        "isOwner": true,
                        "user": {
                          "email": "owner@example.com",
                          "username": "owner"
                        }
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "List all members in the workspace",
        "tags": [
          "Tenant Members"
        ]
      }
    },
    "/tenants/{tenantId}/members/{userId}": {
      "patch": {
        "operationId": "TenantMembersController_updateMember",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Member updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Member updated successfully",
                    "data": {
                      "userId": "user-id",
                      "tenantId": "tenant-id",
                      "isOwner": false,
                      "roleId": "new-role-id"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Update a member role or owner status (Owner only)",
        "tags": [
          "Tenant Members"
        ]
      },
      "delete": {
        "operationId": "TenantMembersController_removeMember",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Member removed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Member removed successfully",
                    "data": null
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Remove a member from the workspace (Owner only)",
        "tags": [
          "Tenant Members"
        ]
      }
    },
...
  },
  "info": {
    "title": "SaaS Multi-Tenant API",
    "description": "Auto-generated OpenAPI specification for the SaaS Demo application.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token to authenticate",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateTenantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The display name of the workspace"
          },
          "slug": {
            "type": "string",
            "description": "Unique URL-friendly slug. Auto-generated if omitted."
          },
          "settings": {
            "type": "object",
            "description": "Optional JSON settings for the workspace"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateTenantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The display name of the workspace"
          },
          "slug": {
            "type": "string",
            "description": "Unique URL-friendly slug. Auto-generated if omitted."
          },
          "settings": {
            "type": "object",
            "description": "Optional JSON settings for the workspace"
          }
        }
      },
      "InviteMemberDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email address of the user to invite"
          },
          "roleId": {
            "type": "string",
            "description": "Role ID to assign to the user upon joining"
          }
        },
        "required": [
          "email"
        ]
      },
      "AcceptInvitationDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "The unique token sent to the user email"
          }
        },
        "required": [
          "token"
        ]
      },
      "UpdateMemberDto": {
        "type": "object",
        "properties": {
          "isOwner": {
            "type": "boolean",
            "description": "Set whether the member is an owner"
          },
          "roleId": {
            "type": "string",
            "description": "Role ID to assign to the member"
          }
        }
      },
...
    }
  }
}